requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
apscheduler==3.10.4
python-multipart==0.0.12
httpx==0.27.0
python-dotenv==1.0.1
weasyprint==62.3
openpyxl==3.1.5
efipay==1.0.4
python-dateutil==2.9.0